deep-learning/image-processing/medical-landmark-detection/requirements.txt
dicompyler-core==0.5.5
pydicom==2.2.2
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
scipy==1.8.1
scikit-image==0.19.3
SimpleITK==2.1.1.1
tqdm==4.51.0
PyYAML==6.0